Fixed after connecting the db

+++ backend/home/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Members(models.Model):
     f_name = models.CharField(max_length=50)
     l_name = models.CharField(max_length=50)
-    phone = models.IntegerField(null=True)
+    phone = models.BigIntegerField(null=True)
     dob = models.DateField(null=True)
     occupation = models.CharField(max_length=50, null=True)
     email = models.CharField(max_length=50, null=True)
